+++ dune/grid/concepts/entity.hh
@@ -76,7 +76,7 @@ namespace Impl {
 template<class E>
 concept EntityExtended = EntityGeneral<E> &&
   Geometry<typename E::LocalGeometry> &&
-requires(const E e, int maxLevel)
+requires(const E e, int maxLevel, int i, unsigned int codim)
 {
   requires (E::codimension == 0);
   { e.father()                   } -> std::same_as<E>;
@@ -89,6 +89,7 @@ requires(const E e, int maxLevel)
   { e.isNew()                    } -> std::convertible_to<bool>;
   { e.mightVanish()              } -> std::convertible_to<bool>;
   { e.hasBoundaryIntersections() } -> std::convertible_to<bool>;
+  { e.partitionType(i,codim)     } -> std::convertible_to<Dune::PartitionType>;
 
   requires std::same_as<E, typename E::template Codim<0>::Entity>;
 } &&